+++ src/components/Map/constants.ts
@@ -2,9 +2,3 @@ export const containerStyle = {
     width: '100%',
     height: '100%',
 };
-
-export const libraries: ('drawing' | 'geometry' | 'places' | 'localContext' | 'visualization')[] = [
-    'drawing',
-    'geometry',
-    'places',
-];
