+++ src/layouts/GoogleMapsLoader/constants.ts
@@ -1,0 +1,5 @@
+export const libraries: ('drawing' | 'geometry' | 'places' | 'localContext' | 'visualization')[] = [
+    'drawing',
+    'geometry',
+    'places',
+];
