+++ src/layouts/GoogleMapsLoader/index.ts
@@ -1,0 +1,3 @@
+import GoogleMapsLoader from './GoogleMapsLoader';
+
+export default GoogleMapsLoader;
