--- src/layouts/PlanLayout/index.ts
@@ -1,3 +1,0 @@
-import PlanLayout from './PlanLayout';
-
-export default PlanLayout;
